map()/filter() return values are no longer lists in python3
In python3, the return values from map() and filter() are special
map/filter objects, and not lists, as in python2. This means that in
order to compare a return value from map() to an expected list, we first
need to "cast" map()'s return to a list, at which point, the whole
expression is probably more readable as a list comprehension, anyway.
Likewise, for filter(), since its return value is no longer a list,
we cannot directly coerce it to boolean in a comparison to check if
the filtered sequence is non-empty. Here, we instead rewrite the
filter() expression to a generator expression, typically wrapped in
any(...) to express the same behavior in a python3-compatible manner.