Coerce dict.keys() return value to a list
In python3, dict.keys() no longer returns a list, so when we want to
compare it to a list of expected values, we need to "cast" it, typically
by using list(). In this case, however, the surrounding pattern (when the
expected list consists of more than one element, we need to _sort_ the
return value from .keys() in order to make it comparable. Therefore, we
use the same sorted(foo.keys()) pattern (which - like list() - returns a
proper list), even when there is only a single expected element.