Add Manifest.from_tar() to create manifests from tar archives
This allows us to directly construct a manifest mapping the contents of a
tar archive, without having to unpack the tar archive into a directory
structure, and then walking that structure.

Include test cases demonstrating correct functionality.

There is an unresolved corner case on python2.6, where an empty file will
incorrectly be interpreted as a proper (but empty) tar(.gz) archive; the
result is that Manifest.from_tar() returns an empty manifest instead of
raising an exception. This is ultimately a bug in python's tarfile
module, but we may want to consider adding a workaround if this (i.e.
feeding empty files to Manifest.from_tar() and expecting an exception)
turns out to be needed in practice.